+++ vowpalwabbit/global_data.h
@@ -23,7 +23,6 @@ namespace po = boost::program_options;
 #include "learner.h"
 #include "allreduce.h"
 #include "v_hashmap.h"
-#include "labeldict.h"
 #include <time.h>
 
 struct version_struct {
@@ -130,6 +129,69 @@ struct dictionary_info {
   feature_dict* dict;
 };
 
+class namedlabels {
+  private:
+
+  v_array<substring> id2name;
+  v_hashmap<substring,uint32_t> name2id;
+  uint32_t K;
+
+  public:
+
+  namedlabels(string label_list) {
+    id2name = v_init<substring>();
+    substring ss = { (char*)label_list.c_str(), nullptr };
+    ss.end = ss.begin + label_list.length();
+    tokenize(',', ss, id2name);
+
+    K = id2name.size();
+    name2id.init(4 * K + 1, 0, substring_equal);
+    for (size_t k=0; k<K; k++) {
+      substring& l = id2name[k];
+      size_t hash = uniform_hash((unsigned char*)l.begin, l.end-l.begin, 378401);
+      uint32_t id = name2id.get(l, hash);
+      if (id != 0) {
+        cerr << "error: label dictionary initialized with multiple occurances of: ";
+        for (char*c=l.begin; c!=l.end; c++) cerr << *c;
+        cerr << endl;
+        throw exception();
+      }
+      size_t len = l.end - l.begin;
+      substring l_copy = { calloc_or_die<char>(len), nullptr };
+      memcpy(l_copy.begin, l.begin, len * sizeof(char));
+      l_copy.end = l_copy.begin + len;
+      name2id.put(l_copy, hash, k+1);
+    }
+  }
+
+  ~namedlabels() {
+    for (size_t i=0; i<id2name.size(); ++i)
+      free(id2name[i].begin);
+    id2name.delete_v();
+  }
+
+  uint32_t getK() { return K; }
+  
+  uint32_t get(substring& s) {
+    size_t hash = uniform_hash((unsigned char*)s.begin, s.end-s.begin, 378401);
+    uint32_t v  =  name2id.get(s, hash);
+    if (v == 0) {
+      cerr << "warning: missing named label '";
+      for (char*c = s.begin; c != s.end; c++) cerr << *c;
+      cerr << '\'' << endl;
+    }
+    return v;
+  }
+
+  substring get(uint32_t v) {
+    if ((v == 0) || (v > K)) {
+      substring ss = {nullptr,nullptr};
+      return ss;
+    } else
+      return id2name[v-1];
+  }
+};
+
 struct shared_data {
   size_t queries;
 
@@ -149,7 +211,7 @@ struct shared_data {
   float min_label;//minimum label encountered
   float max_label;//maximum label encountered
 
-  labeldict* ldict;
+  namedlabels* ldict;
   
   //for holdout
   double weighted_holdout_examples;
